Cloud, Seeing, Downtime renamed and decoupled. TimeHandler moved into sims_utils
