+++ python/lsst/sims/ocs/environment/__init__.py
@@ -1,5 +1,5 @@
 """
 Module for classes dealing with environmental conditions.
 """
-from .cloud_model import *
-from .seeing_model import *
+from .cloud_interface import *
+from .seeing_interface import *
