+++ ups/sims_ocs.table
@@ -5,6 +5,9 @@ setupRequired(sconsUtils)
 setupRequired(sims_survey_fields)
 setupRequired(ts_astrosky_model)
 setupRequired(ts_observatory_model)
+setupRequired(sims_seeingModel)
+setupRequired(sims_cloudModel)
+setupRequired(sims_downtimeModel)
 
 envPrepend(PYTHONPATH, ${PRODUCT_DIR}/python)
 envPrepend(PATH, ${PRODUCT_DIR}/scripts)
